backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.5
transformers==4.35.2
torch==2.1.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
reportlab==4.0.7
aiofiles==23.2.1
email-validator==2.1.0
python-magic==0.4.27
Pillow==10.1.0
python-docx==1.1.0
jinja2==3.1.2
markdown==3.5.1
bleach==6.1.0
cryptography==41.0.8
bcrypt==4.1.2